requirements.txt
numpy
scipy
pandas
scikit-learn
tensorflow
matplotlib
seaborn
pymanopt
pyyaml